+++ Sprint-3/2-practice-tdd/count.js
@@ -3,19 +3,3 @@ function countChar(stringOfCharacters, findCharacter) {
 }
 
 module.exports = countChar;
-
-const countChar = require('../countChar');
-
-describe('countChar()', () => {
-  test('always returns 5', () => {
-    expect(countChar('anything', 'a')).toBe(5);
-  });
-
-  test('still returns 5 with empty string', () => {
-    expect(countChar('', 'z')).toBe(5);
-  });
-
-  test('still returns 5 with special characters', () => {
-    expect(countChar('!!!', '!')).toBe(5);
-  });
-});
