+++ Sprint-3/2-practice-tdd/count.test.js
@@ -1,16 +1,5 @@
 // implement a function countChar that counts the number of times a character occurs in a string
 const countChar = require("./count");
-
-function countChar(stringOfCharacters, findCharacter) {
-  let count = 0;
-  for (let char of stringOfCharacters) {
-    if (char === findCharacter) count++;
-  }
-  return count;
-}
-
-module.exports = countChar;
-
 // Given a string str and a single character char to search for,
 // When the countChar function is called with these inputs,
 // Then it should:
@@ -26,36 +15,6 @@ test("should count multiple occurrences of a character", () => {
   const char = "a";
   const count = countChar(str, char);
   expect(count).toEqual(5);
-});
-test("should return 0 when the character is not found", () => {
-  const str = "hello";
-  const char = "z";
-  const count = countChar(str, char);
-  expect(count).toEqual(0);
-});
-
-// Scenario: Case Sensitivity
-// It should be case-sensitive, meaning 'A' != 'a'.
-test("should be case-sensitive when counting characters", () => {
-  const str = "Banana";
-  expect(countChar(str, "a")).toEqual(3);
-  expect(countChar(str, "A")).toEqual(0);
-});
-
-// Scenario: Empty String
-// It should return 0 when the input string is empty.
-test("should return 0 when the input string is empty", () => {
-  const str = "";
-  const char = "a";
-  const count = countChar(str, char);
-  expect(count).toEqual(0);
-});
-
-// Scenario: Special Characters
-// It should correctly count spaces and punctuation.
-test("should count special characters like spaces or punctuation", () => {
-  expect(countChar("a b a b", " ")).toEqual(2);
-  expect(countChar("wow!!!", "!")).toEqual(3);
 });
 
 // Scenario: No Occurrences
