--- Sprint-3/2-practice-tdd/repeat-str.js
@@ -1,5 +1,0 @@
-function repeatStr() {
-  return "hellohellohello";
-}
-
-module.exports = repeatStr;
