+++ Sprint-3/2-practice-tdd/repeat.js
@@ -1,11 +1,5 @@
-const repeat = require("./repeat");
-
 function repeat() {
   return "hellohellohello";
 }
 
 module.exports = repeat;
-
-test("should repeat the given word three times", () => {
-  expect(repeat("hi", 3)).toBe("hihihi");
-});
